+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_box_close_v2.py
@@ -40,7 +40,10 @@ class SawyerBoxCloseEnvV2(SawyerXYZEnv):
 
         self._target_to_obj_init = None
 
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
         self._random_reset_space = Box(
             np.hstack((obj_low, goal_low)),
             np.hstack((obj_high, goal_high)),
