+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_hammer_v2.py
@@ -41,7 +41,10 @@ class SawyerHammerEnvV2(SawyerXYZEnv):
         self.nail_init_pos = None
 
         self._random_reset_space = Box(np.array(obj_low), np.array(obj_high))
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
 
     @property
     def model_name(self):
