+++ metaworld/envs/mujoco/sawyer_xyz/v2/sawyer_push_wall_v2.py
@@ -63,7 +63,10 @@ class SawyerPushWallEnvV2(SawyerXYZEnv):
             np.hstack((obj_low, goal_low)),
             np.hstack((obj_high, goal_high)),
         )
-        self.goal_space = Box(np.array(goal_low), np.array(goal_high))
+        self.goal_space = Box(
+            np.array(goal_low, dtype=np.float32),
+            np.array(goal_high, dtype=np.float32),
+        )
 
         self.num_resets = 0
